+++ pypls/__init__.py
@@ -1,0 +1,6 @@
+from .cross_validation import CrossValidation
+
+
+__all__ = [
+    'CrossValidation'
+]
